backend/app/services/ocr_service.py
import os
from PIL import Image
import io
import requests
from flask import current_app

class OCRService:
    """
    Service for OCR processing of scanned test papers
    """
    
    @staticmethod
    def process_image(image_data, document_type='test_submission'):
        """
        Process scanned image using OCR
        
        Args:
            image_data (bytes): Binary image data
            document_type (str): Type of document ('test_submission', 'answer_sheet')
            
        Returns:
            dict: Extracted text with position information
        """
        # This is a placeholder for actual OCR API integration
        # In production, this would call Google Document AI or Azure Document Intelligence
        
        # Example of how to use Google Document AI (commented out)
        if 'GOOGLE_APPLICATION_CREDENTIALS' in os.environ:
            # Simulated API request
            # In production, would use the actual Document AI client library
            return OCRService._mock_google_document_ai_response(document_type)
        
        # Example of how to use Azure Document Intelligence (commented out)
        if 'AZURE_FORM_RECOGNIZER_KEY' in os.environ and 'AZURE_FORM_RECOGNIZER_ENDPOINT' in os.environ:
            # Simulated API request
            # In production, would use the actual Azure SDK
            return OCRService._mock_azure_document_intelligence_response(document_type)
        
        # If no credentials are available, use a mock response for development
        return OCRService._mock_ocr_response(document_type)
    
    @staticmethod
    def _mock_ocr_response(document_type):
        """
        Generate a mock OCR response for development
        """
        if document_type == 'test_submission':
            return {
                'pages': [
                    {
                        'page_number': 1,
                        'dimensions': {'width': 8.5, 'height': 11, 'unit': 'inch'},
                        'text_blocks': [
                            {
                                'id': 'block1',
                                'text': 'Student ID: 12345',
                                'confidence': 0.95,
                                'bounding_box': {
                                    'vertices': [
                                        {'x': 1, 'y': 1},
                                        {'x': 3, 'y': 1},
                                        {'x': 3, 'y': 1.5},
                                        {'x': 1, 'y': 1.5}
                                    ]
                                }
                            },
                            {
                                'id': 'block2',
                                'text': 'Question 1: Sample answer written by student for demonstration purposes.',
                                'confidence': 0.92,
                                'bounding_box': {
                                    'vertices': [
                                        {'x': 1, 'y': 2},
                                        {'x': 7, 'y': 2},
                                        {'x': 7, 'y': 4},
                                        {'x': 1, 'y': 4}
                                    ]
                                }
                            }
                        ]
                    }
                ],
                'document_metadata': {
                    'test_id': '123',
                    'student_id': '12345'
                }
            }
        else:
            return {
                'pages': [
                    {
                        'page_number': 1,
                        'text_blocks': [
                            {
                                'id': 'block1',
                                'text': 'Sample extracted text from document',
                                'confidence': 0.9
                            }
                        ]
                    }
                ]
            }
    
    @staticmethod
    def _mock_google_document_ai_response(document_type):
        """
        Generate a mock response similar to Google Document AI
        """
        return {
            'document': {
                'pages': [
                    {
                        'pageNumber': 1,
                        'dimension': {'width': 8.5, 'height': 11, 'unit': 'inch'},
                        'blocks': [
                            {
                                'blockType': 'TEXT',
                                'content': 'Student answer to question 1: The process of photosynthesis involves...',
                                'confidence': 0.95
                            }
                        ]
                    }
                ],
                'entities': [
                    {
                        'type': 'QUESTION_NUMBER',
                        'mentionText': '1',
                        'confidence': 0.98
                    },
                    {
                        'type': 'STUDENT_ID',
                        'mentionText': '12345',
                        'confidence': 0.99
                    }
                ]
            }
        }
    
    @staticmethod
    def _mock_azure_document_intelligence_response(document_type):
        """
        Generate a mock response similar to Azure Document Intelligence
        """
        return {
            'status': 'succeeded',
            'analyzeResult': {
                'pages': [
                    {
                        'pageNumber': 1,
                        'width': 8.5,
                        'height': 11,
                        'unit': 'inch',
                        'lines': [
                            {
                                'content': 'Student ID: 12345',
                                'boundingBox': [1, 1, 3, 1, 3, 1.5, 1, 1.5],
                                'confidence': 0.95
                            },
                            {
                                'content': 'Question 1: Sample student answer text.',
                                'boundingBox': [1, 2, 7, 2, 7, 4, 1, 4],
                                'confidence': 0.92
                            }
                        ]
                    }
                ],
                'documents': [
                    {
                        'docType': 'test_submission',
                        'fields': {
                            'studentId': {'content': '12345', 'confidence': 0.99},
                            'testId': {'content': '123', 'confidence': 0.98}
                        }
                    }
                ]
            }
        } 