backend/requirements.txt
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
flask-cors==5.0.1
openai==1.70.0
psycopg2-binary==2.9.10
pymongo==4.11.3
python-dotenv==1.1.0
pydantic==2.11.2
pillow==11.1.0
requests==2.32.3
gunicorn==23.0.0
pytest==8.3.5
black==25.1.0
flake8==7.2.0
